New class created to separate pre-processing from run class
